requirements.txt
Flask==2.3.3
pytesseract==0.3.8
Pillow==10.2.0
Werkzeug==2.3.7
python-dotenv==0.19.0
setuptools==69.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyPDF2==3.0.1
python-magic-bin==0.4.14
